+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,6 +19,8 @@
 -->
 
 ## Upcoming Release
+* Upgraded Azure.Identity to 1.11.2 and Azure.Identity.Broker to 1.1.0.
+* Upgraded Azure.Core to 1.38.0 and Microsoft.Identity.Client to 4.60.3.
 * Supported interactive subscription selection for user login flow.
 * Added config `LoginExperienceV2` to allow customer to switch the default behavior of context selection back.
 * Updated the reference of Azure PowerShell Common to 1.3.94-preview.